+++ R/crtree.R
@@ -74,8 +74,10 @@ crtree <- function(dataset, rvar, evar, type = "", lev = "", wts = "None", minsp
     }
   }
 
-  if (any(summarise_all(dataset, does_vary) == FALSE)) {
-    return("One or more selected variables show no variation. Please select other variables." %>% add_class("crtree"))
+  not_vary <- colnames(dataset)[summarise_all(dataset, does_vary) == FALSE]
+  if (length(not_vary) > 0) {
+    return(paste0("The following variable(s) show no variation. Please select other variables.\n\n** ", paste0(not_vary, collapse = ", "), " **") %>%
+      add_class("crtree"))
   }
 
   rv <- dataset[[rvar]]