+++ R/nn.R
@@ -70,8 +70,10 @@ nn <- function(
     }
   }
 
-  if (any(summarise_all(dataset, .funs = does_vary) == FALSE)) {
-    return("One or more selected variables show no variation. Please select other variables." %>% add_class("nn"))
+  not_vary <- colnames(dataset)[summarise_all(dataset, does_vary) == FALSE]
+  if (length(not_vary) > 0) {
+    return(paste0("The following variable(s) show no variation. Please select other variables.\n\n** ", paste0(not_vary, collapse = ", "), " **") %>%
+      add_class("nn"))
   }
 
   rv <- dataset[[rvar]]